# Ensure hashes are independent of endianness