Closes #1084 

@QuLogic do the tests pass without modification on big-endian systems now?

@easyaspi314 if you get a moment, could you please comment on whether this looks appropriate as a way to ensure that the resulting hash is independent of endianness? Thanks!