@easyaspi314 2021-01-05
Yes, that looks correct.