@QuLogic 2021-01-08
Sorry, unfortunately that is not enough. The problem is that the serialization itself is not endian-independent. From the comment, I assumed that was intentional:
https://github.com/r-lib/rlang/blob/316ee6860bd41bb23d40af430aa396168bedfe46/src/internal/hash.c#L113-L116