@DavisVaughan 2021-01-25
So, it is somewhat complicated to implement endian independent hashes.

For serialization version 3, we need to decode the `int` corresponding to the native encoding size. This would require the use of `R_XDRDecodeInteger()`, which I started to use but then realized it is a hidden function that shouldn't be used. To work around this, I took a look at its implementation and saw that it is really just a call to `ntohl()` to do a byte swap when required, so I've re-implemented that here. I think this is a standard function on most OSes (maybe?), but I'm using the version that R uses.

The good news is that, on my little-endian Mac, we now get the hash values that were reported for the big endian system in https://github.com/r-lib/rlang/pull/1084/files

The bad news is that it does have a pretty big impact on performance.

``` r
library(rlang)

x <- 1:1e6 + 0L
y <- as.list(x)

bench::mark(
  hash(x),
  hash(y),
  check = FALSE,
  iterations = 50
)

# Before
#> # A tibble: 2 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 hash(x)     198.6µs  202.7µs    4699.    14.95KB        0
#> 2 hash(y)      47.3ms   49.7ms      20.0    8.63KB        0

# After
#> # A tibble: 2 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 hash(x)      6.24ms   6.52ms     153.    14.95KB        0
#> 2 hash(y)     82.41ms  86.48ms      11.5    8.63KB        0
```

The other bad news is that this obviously adds some complexity.

We could merge this, or we could state that the hashes are guaranteed to be reproducible within the same endianness.