@lionel- 2021-01-26
I think you're right, ensuring stability across endianness is not worth the added complexity and the performance penalty.