@lionel- 2021-01-26
Though maybe check the approach in https://commandcenter.blogspot.com/2012/04/byte-order-fallacy.html