@DavisVaughan 2021-01-26
That actually worked great!

So now the main question is one of performance, since the implementation is much easier to explain now.

I am also using `R_XDR_INTEGER_SIZE`, a define that is defined in `Rinternals.h` and is always set to `4` since they use an `int32_t` when doing XDR encoding/decoding. It is right next to a group of functions that have all been "hidden" on Windows (like the `R_XDRDecodeInteger()` helper I was going to use). Is it safe to use this define or should I just replace it with a 4 here?
https://github.com/wch/r-source/blob/a1425adea54bcc98eef86081522b5dbb3e149cdc/src/include/Rinternals.h#L1358-L1364